feat: scaffold Vue app structure with component wrappers
