+++ index.html
@@ -141,7 +141,9 @@
     </button>
   </nav>
 
+  <div id="app"></div>
   <!-- Load the modular JavaScript -->
+  <script src="https://unpkg.com/vue@3/dist/vue.global.prod.js"></script>
   <script type="module" src="main.js"></script>
 </body>
