+++ src/components/Audio.js
@@ -1,0 +1,12 @@
+import { initAudio, initAudioUI } from '../../modules/audio.js';
+
+export default {
+  name: 'AudioPlayer',
+  mounted() {
+    initAudio();
+    initAudioUI();
+  },
+  render() {
+    return null;
+  }
+};
