+++ src/components/Gallery.js
@@ -1,0 +1,11 @@
+import { initGallery } from '../../modules/gallery.js';
+
+export default {
+  name: 'Gallery',
+  mounted() {
+    initGallery();
+  },
+  render() {
+    return null;
+  }
+};
