+++ src/components/Sidebar.js
@@ -1,0 +1,11 @@
+import { initSidebar } from '../../modules/sidebar.js';
+
+export default {
+  name: 'Sidebar',
+  mounted() {
+    initSidebar();
+  },
+  render() {
+    return null;
+  }
+};
