+++ src/components/TagExplorer.js
@@ -1,0 +1,9 @@
+import { openTagExplorer } from '../../modules/tag-explorer.js';
+
+export default {
+  name: 'TagExplorer',
+  methods: { openTagExplorer },
+  render() {
+    return null;
+  }
+};
